[MIN-31] updated pallet controller to Substrate 3.0.0

+++ modules/accounts/src/mock.rs
@@ -56,7 +56,7 @@ impl system::Config for Test {
 }
 
 parameter_types! {
-	pub const MaxMembers: u32 = 16;
+	pub const MaxMembers: u8 = 16;
 }
 
 impl test_accounts::Config for Test {
